v9 reachability: anti-CGNAT punch cascade + LowID-agnostic eD2K + relay hardening
This session's work. All cap-gated + default-OFF (backward-compat with vanilla 0.70b
and prior forks — wire is byte-identical until a pref is flipped); each touched TU
compile-validated; pending multi-PC validation.

Anti-CGNAT
- Port-prediction ("birthday spray"): SendEseHolePunchReqSpray fires the 2-way punch at
  a clamped window (+/-[0,8]) of ports around the observed one so a symmetric-NAT/CGNAT
  per-destination external port can be hit. Pref EseHolePunchPortPredict. Applies to both
  eD2K (TryToConnect) and Live (selector) via the shared primitive.
- R.3 buddy relay made discoverable + safe: advertise ESE_CAP_LIVE_RELAY (bit 21) iff
  EseRelayAccept; hard egress byte-budget (EseRelayMaxKBps, rolling 1s window, effective
  bytes = len*(1+downstream)) sheds load over budget; broadcaster only picks
  RDV-relay-capable buddies. Closes the free-CDN abuse vector.

LowID-agnostic
- eD2K downloads escalate Direct -> punch2 -> punch3: on symmetric-NAT give-up,
  TryToConnect fires one 3-way rendezvous (InitiateKad3Rendezvous) via an RDV-capable
  connected peer, gated on the target advertising the RDV cap. Pref EseEd2kPunch3,
  one-shot per peer (m_bNatRdvTried, reset on uTP accept). Reuses the Live-validated R.1
  machinery; no eD2K data relay.

Observability / control
- Reach-cascade outcome telemetry: which stage (Direct/Punch2/Punch3/Relay) actually
  connected each source (was collapsed into REACH_DONE).
- /api/ese/v9?on=1 single lever flips all v9 reachability prefs + keepalive for
  one-command multi-PC bring-up; on=0 = soft revert to vanilla reachability.
- Counters in /api/ese/v9, /api/relay, /api/holepunch/test; NetworkInfo decodes RDV+Relay.

Cleanup
- Removed dead CHolePuncher (skeleton with TODO sends, 0 callers, superseded by the live
  R.1 rendezvous in CKademliaUDPListener).

Pre-commit adversarial review (6 reviewers + verify): 0 critical/major, 1 minor fixed.

Also consolidates accumulated WIP: docs reorg (INDEX + archive/ + protocol registry +
linter), libreach reference library (L1-L3 + tests), NodeIdentity / PartHashThread /
HolePunchCookie subsystems, ese-server JS. libutp submodule change intentionally excluded.

+++ srchybrid/AICHSyncThread.cpp
@@ -67,7 +67,7 @@ int CAICHSyncThread::Run()
 			return 0;
 		}
 	}
-	uint32 nLastVerifiedPos = 0;
+	ULONGLONG nLastVerifiedPos = 0;
 	try {
 		if (file.GetLength() >= 1) {
 			uint8 header = file.ReadUInt8();
@@ -85,7 +85,7 @@ int CAICHSyncThread::Run()
 
 				// skip the rest of this hashset
 				file.Seek(nHashCount * (LONGLONG)CAICHHash::GetHashSize(), CFile::current);
-				nLastVerifiedPos = (uint32)file.GetPosition();
+				nLastVerifiedPos = file.GetPosition();
 			}
 		} else
 			file.WriteUInt8(KNOWN2_MET_VERSION);
